algunas funciones privadas y diagramas en Imagenes

+++ Memoria.java
@@ -37,7 +37,7 @@ public class Memoria {
             }
         }
             try {
-                Thread.sleep(((Math.random() * 200) + 50)); //se demora entre alfa 50ms y beta 200ms. Bastara con esto para cumplir con el enunciado?
+                Thread.sleep((long)((Math.random() * 200) + 50)); //se demora entre alfa 50ms y beta 200ms. Bastara con esto para cumplir con el enunciado?
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
